Create README.md
This code defines a function is_prime that checks if a given number is prime. It iterates through numbers up to the square root of the input number and checks if any of them divide evenly into the input number. If none do, the number is prime. This is a useful utility function that can be used in various scenarios where prime numbers are involved, such as cryptography, algorithms, or mathematical computations.






Signed-off-by: Victor Keenom <victorkeenom1234@gmail.com>